+++ lite/core/CMakeLists.txt
@@ -125,3 +125,5 @@ lite_cc_test (test_type_system SRCS type_system_test.cc)
 lite_cc_test (test_types SRCS types_test.cc)
 lite_cc_test (test_memory SRCS memory_test.cc)
 lite_cc_test (test_context SRCS context_test.cc)
+lite_cc_test(test_scalar SRCS scalar_test.cc)
+lite_cc_test(test_int_array SRCS int_array_test.cc)
