+++ lite/core/int_array_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "lite/operators/int_array.h"
+#include "lite/core/int_array.h"
 #include <gtest/gtest.h>
 #include <vector>
