+++ lite/core/scalar.h
